@@ -118,7 +118,10 @@ enum MomentSkippedReason {
   tap_outside('tap_outside'),
 
   /// issuing_failed
-  issuing_failed('issuing_failed');
+  issuing_failed('issuing_failed'),
+
+  /// Unknown reason
+  unknown_reason('unknown_reason');
 
   ///
   const MomentSkippedReason(String reason) : _reason = reason;
@@ -137,7 +140,10 @@ enum MomentDismissedReason {
   cancel_called('cancel_called'),
 
   /// flow_restarted
-  flow_restarted('flow_restarted');
+  flow_restarted('flow_restarted'),
+
+  /// Unknown reason
+  unknown_reason('unknown_reason');
 
   ///
   const MomentDismissedReason(String reason) : _reason = reason;
@@ -207,7 +213,10 @@ enum CredentialSelectBy {
   btn_confirm_add_session('btn_confirm_add_session'),
 
   /// A user with an existing session used the browser's "FedCM" flow.
-  fedcm('fedcm');
+  fedcm('fedcm'),
+
+  /// A fedcm authentication without user intervention.
+  fedcm_auto('fedcm_auto');
 
   ///
   const CredentialSelectBy(String selectBy) : _selectBy = selectBy;
