@@ -15,6 +15,7 @@ import 'package:http/http.dart' as http;
 import 'src/sign_in_button.dart';
 
 /// The scopes required by this application.
+// #docregion Initialize
 const List<String> scopes = <String>[
   'email',
   'https://www.googleapis.com/auth/contacts.readonly',
@@ -25,6 +26,7 @@ GoogleSignIn _googleSignIn = GoogleSignIn(
   // clientId: 'your-client_id.apps.googleusercontent.com',
   scopes: scopes,
 );
+// #enddocregion Initialize
 
 void main() {
   runApp(
@@ -55,12 +57,14 @@ class _SignInDemoState extends State<SignInDemo> {
 
     _googleSignIn.onCurrentUserChanged
         .listen((GoogleSignInAccount? account) async {
+// #docregion CanAccessScopes
       // In mobile, being authenticated means being authorized...
       bool isAuthorized = account != null;
-      // However, in the web...
+      // However, on web...
       if (kIsWeb && account != null) {
         isAuthorized = await _googleSignIn.canAccessScopes(scopes);
       }
+// #enddocregion CanAccessScopes
 
       setState(() {
         _currentUser = account;
@@ -136,6 +140,7 @@ class _SignInDemoState extends State<SignInDemo> {
   //
   // On the web, the on-click handler of the Sign In button is owned by the JS
   // SDK, so this method can be considered mobile only.
+  // #docregion SignIn
   Future<void> _handleSignIn() async {
     try {
       await _googleSignIn.signIn();
@@ -143,6 +148,7 @@ class _SignInDemoState extends State<SignInDemo> {
       print(error);
     }
   }
+  // #enddocregion SignIn
 
   // Prompts the user to authorize `scopes`.
   //
@@ -150,14 +156,18 @@ class _SignInDemoState extends State<SignInDemo> {
   // and Authorization at the same time (like the web).
   //
   // On the web, this must be called from an user interaction (button click).
+  // #docregion RequestScopes
   Future<void> _handleAuthorizeScopes() async {
     final bool isAuthorized = await _googleSignIn.requestScopes(scopes);
+    // #enddocregion RequestScopes
     setState(() {
       _isAuthorized = isAuthorized;
     });
+    // #docregion RequestScopes
     if (isAuthorized) {
       unawaited(_handleGetContact(_currentUser!));
     }
+    // #enddocregion RequestScopes
   }
 
   Future<void> _handleSignOut() => _googleSignIn.disconnect();
