@@ -161,6 +161,9 @@ Page resource error:
           })
           ..setOnUrlChange((UrlChange change) {
             debugPrint('url change to ${change.url}');
+          })
+          ..setOnHttpAuthRequest((HttpAuthRequest request) {
+            openDialog(request);
           }),
       )
       ..addJavaScriptChannel(JavaScriptChannelParams(
@@ -179,9 +182,11 @@ Page resource error:
           request.grant();
         },
       )
-      ..loadRequest(LoadRequestParams(
-        uri: Uri.parse('https://flutter.dev'),
-      ));
+      ..loadRequest(
+        LoadRequestParams(
+          uri: Uri.parse('https://flutter.dev'),
+        ),
+      );
   }
 
   @override
@@ -219,6 +224,50 @@ Page resource error:
       child: const Icon(Icons.favorite),
     );
   }
+
+  Future<void> openDialog(HttpAuthRequest httpRequest) async {
+    final TextEditingController usernameTextController =
+        TextEditingController();
+    final TextEditingController passwordTextController =
+        TextEditingController();
+
+    return showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('${httpRequest.host}: ${httpRequest.realm ?? '-'}'),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              TextField(
+                decoration: const InputDecoration(labelText: 'Username'),
+                autofocus: true,
+                controller: usernameTextController,
+              ),
+              TextField(
+                decoration: const InputDecoration(labelText: 'Password'),
+                controller: passwordTextController,
+              ),
+            ],
+          ),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () {
+                httpRequest.onProceed(
+                  WebViewCredential(
+                    user: usernameTextController.text,
+                    password: passwordTextController.text,
+                  ),
+                );
+                Navigator.of(context).pop();
+              },
+              child: const Text('Authenticate'),
+            ),
+          ],
+        );
+      },
+    );
+  }
 }
 
 enum MenuOptions {
@@ -237,6 +286,7 @@ enum MenuOptions {
   setCookie,
   videoExample,
   logExample,
+  basicAuthentication,
 }
 
 class SampleMenu extends StatelessWidget {
@@ -303,6 +353,9 @@ class SampleMenu extends StatelessWidget {
           case MenuOptions.logExample:
             _onLogExample();
             break;
+          case MenuOptions.basicAuthentication:
+            _promptForUrl(context);
+            break;
         }
       },
       itemBuilder: (BuildContext context) => <PopupMenuItem<MenuOptions>>[
@@ -366,6 +419,10 @@ class SampleMenu extends StatelessWidget {
         const PopupMenuItem<MenuOptions>(
           value: MenuOptions.videoExample,
           child: Text('Video example'),
+        ),
+        const PopupMenuItem<MenuOptions>(
+          value: MenuOptions.basicAuthentication,
+          child: Text('Basic Authentication Example'),
         ),
       ],
     );
@@ -546,8 +603,41 @@ class SampleMenu extends StatelessWidget {
       debugPrint(
           '== JS == ${consoleMessage.level.name}: ${consoleMessage.message}');
     });
-
     return webViewController.loadHtmlString(kLogExamplePage);
+  }
+
+  Future<void> _promptForUrl(BuildContext context) {
+    final TextEditingController urlTextController = TextEditingController();
+
+    return showDialog<String>(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Input URL to visit'),
+          content: TextField(
+            decoration: const InputDecoration(labelText: 'URL'),
+            autofocus: true,
+            controller: urlTextController,
+          ),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () {
+                if (urlTextController.text.isNotEmpty) {
+                  final Uri? uri = Uri.tryParse(urlTextController.text);
+                  if (uri != null && uri.scheme.isNotEmpty) {
+                    webViewController.loadRequest(
+                      LoadRequestParams(uri: uri),
+                    );
+                    Navigator.pop(context);
+                  }
+                }
+              },
+              child: const Text('Visit'),
+            ),
+          ],
+        );
+      },
+    );
   }
 }
 
