@@ -2,10 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:typed_data';
 import 'dart:ui';
 
-import 'package:flutter/services.dart' show BinaryMessenger;
+import 'package:flutter/services.dart' show BinaryMessenger, Uint8List;
 
 import 'android_webview.dart';
 import 'android_webview.g.dart';
@@ -50,6 +49,7 @@ class AndroidWebViewFlutterApis {
     PermissionRequestFlutterApiImpl? permissionRequestFlutterApi,
     CustomViewCallbackFlutterApiImpl? customViewCallbackFlutterApi,
     ViewFlutterApiImpl? viewFlutterApi,
+    HttpAuthHandlerFlutterApiImpl? httpAuthHandlerFlutterApi,
   }) {
     this.javaObjectFlutterApi =
         javaObjectFlutterApi ?? JavaObjectFlutterApiImpl();
@@ -72,6 +72,8 @@ class AndroidWebViewFlutterApis {
     this.customViewCallbackFlutterApi =
         customViewCallbackFlutterApi ?? CustomViewCallbackFlutterApiImpl();
     this.viewFlutterApi = viewFlutterApi ?? ViewFlutterApiImpl();
+    this.httpAuthHandlerFlutterApi =
+        httpAuthHandlerFlutterApi ?? HttpAuthHandlerFlutterApiImpl();
   }
 
   static bool _haveBeenSetUp = false;
@@ -115,6 +117,9 @@ class AndroidWebViewFlutterApis {
   /// Flutter Api for [View].
   late final ViewFlutterApiImpl viewFlutterApi;
 
+  /// Flutter Api for [HttpAuthHandler].
+  late final HttpAuthHandlerFlutterApiImpl httpAuthHandlerFlutterApi;
+
   /// Ensures all the Flutter APIs have been setup to receive calls from native code.
   void ensureSetUp() {
     if (!_haveBeenSetUp) {
@@ -130,6 +135,7 @@ class AndroidWebViewFlutterApis {
       PermissionRequestFlutterApi.setup(permissionRequestFlutterApi);
       CustomViewCallbackFlutterApi.setup(customViewCallbackFlutterApi);
       ViewFlutterApi.setup(viewFlutterApi);
+      HttpAuthHandlerFlutterApi.setup(httpAuthHandlerFlutterApi);
       _haveBeenSetUp = true;
     }
   }
@@ -803,14 +809,47 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     if (instance!.doUpdateVisitedHistory != null) {
       instance.doUpdateVisitedHistory!(webViewInstance!, url, isReload);
+    }
+  }
+
+  @override
+  void onReceivedHttpAuthRequest(
+    int instanceId,
+    int webViewInstanceId,
+    int httpAuthHandlerInstanceId,
+    String host,
+    String realm,
+  ) {
+    final WebViewClient? instance = instanceManager
+        .getInstanceWithWeakReference(instanceId) as WebViewClient?;
+    final WebView? webViewInstance = instanceManager
+        .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
+    final HttpAuthHandler? httpAuthHandlerInstance =
+        instanceManager.getInstanceWithWeakReference(httpAuthHandlerInstanceId)
+            as HttpAuthHandler?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
+    );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
+    );
+    assert(
+      httpAuthHandlerInstance != null,
+      'InstanceManager does not contain a HttpAuthHandler with instanceId: $httpAuthHandlerInstanceId',
+    );
+    if (instance!.onReceivedHttpAuthRequest != null) {
+      return instance.onReceivedHttpAuthRequest!(
+          webViewInstance!, httpAuthHandlerInstance!, host, realm);
     }
   }
 }
@@ -1387,6 +1426,71 @@ class CookieManagerHostApiImpl extends CookieManagerHostApi {
       instanceManager.getIdentifier(instance)!,
       instanceManager.getIdentifier(webView)!,
       accept,
+    );
+  }
+}
+
+/// Host api implementation for [HttpAuthHandler].
+class HttpAuthHandlerHostApiImpl extends HttpAuthHandlerHostApi {
+  /// Constructs a [HttpAuthHandlerHostApiImpl].
+  HttpAuthHandlerHostApiImpl({
+    super.binaryMessenger,
+    InstanceManager? instanceManager,
+  }) : _instanceManager = instanceManager ?? JavaObject.globalInstanceManager;
+
+  /// Maintains instances stored to communicate with native language objects.
+  final InstanceManager _instanceManager;
+
+  /// Helper method to convert instance ids to objects.
+  Future<void> cancelFromInstance(HttpAuthHandler instance) {
+    return cancel(_instanceManager.getIdentifier(instance)!);
+  }
+
+  /// Helper method to convert instance ids to objects.
+  Future<void> proceedFromInstance(
+    HttpAuthHandler instance,
+    String username,
+    String password,
+  ) {
+    return proceed(
+      _instanceManager.getIdentifier(instance)!,
+      username,
+      password,
+    );
+  }
+
+  /// Helper method to convert instance ids to objects.
+  Future<bool> useHttpAuthUsernamePasswordFromInstance(
+    HttpAuthHandler instance,
+  ) {
+    return useHttpAuthUsernamePassword(
+      _instanceManager.getIdentifier(instance)!,
+    );
+  }
+}
+
+/// Flutter API implementation for [HttpAuthHandler].
+class HttpAuthHandlerFlutterApiImpl extends HttpAuthHandlerFlutterApi {
+  /// Constructs a [HttpAuthHandlerFlutterApiImpl].
+  HttpAuthHandlerFlutterApiImpl({
+    this.binaryMessenger,
+    InstanceManager? instanceManager,
+  }) : instanceManager = instanceManager ?? JavaObject.globalInstanceManager;
+
+  /// Receives binary data across the Flutter platform barrier.
+  ///
+  /// If it is null, the default BinaryMessenger will be used which routes to
+  /// the host platform.
+  final BinaryMessenger? binaryMessenger;
+
+  /// Maintains instances stored to communicate with native language objects.
+  final InstanceManager instanceManager;
+
+  @override
+  void create(int instanceId) {
+    instanceManager.addHostCreatedInstance(
+      HttpAuthHandler(),
+      instanceId,
     );
   }
 }
