@@ -256,7 +256,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
           Invocation.getter(#androidWebChromeClient),
         ),
       ) as _i2.WebChromeClient);
-
   @override
   _i2.WebViewClient get androidWebViewClient => (super.noSuchMethod(
         Invocation.getter(#androidWebViewClient),
@@ -269,7 +268,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
           Invocation.getter(#androidWebViewClient),
         ),
       ) as _i2.WebViewClient);
-
   @override
   _i2.DownloadListener get androidDownloadListener => (super.noSuchMethod(
         Invocation.getter(#androidDownloadListener),
@@ -282,7 +280,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
           Invocation.getter(#androidDownloadListener),
         ),
       ) as _i2.DownloadListener);
-
   @override
   _i3.PlatformNavigationDelegateCreationParams get params =>
       (super.noSuchMethod(
@@ -297,7 +294,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
           Invocation.getter(#params),
         ),
       ) as _i3.PlatformNavigationDelegateCreationParams);
-
   @override
   _i9.Future<void> setOnLoadRequest(_i8.LoadRequestCallback? onLoadRequest) =>
       (super.noSuchMethod(
@@ -308,7 +304,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnNavigationRequest(
           _i3.NavigationRequestCallback? onNavigationRequest) =>
@@ -320,7 +315,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnPageStarted(_i3.PageEventCallback? onPageStarted) =>
       (super.noSuchMethod(
@@ -331,7 +325,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnPageFinished(_i3.PageEventCallback? onPageFinished) =>
       (super.noSuchMethod(
@@ -342,7 +335,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnProgress(_i3.ProgressCallback? onProgress) =>
       (super.noSuchMethod(
@@ -353,7 +345,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnWebResourceError(
           _i3.WebResourceErrorCallback? onWebResourceError) =>
@@ -365,7 +356,6 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnUrlChange(_i3.UrlChangeCallback? onUrlChange) =>
       (super.noSuchMethod(
@@ -376,7 +366,17 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
+  @override
+  _i9.Future<void> setOnHttpAuthRequest(
+          _i3.HttpAuthRequestCallback? onHttpAuthRequest) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #setOnHttpAuthRequest,
+          [onHttpAuthRequest],
+        ),
+        returnValue: _i9.Future<void>.value(),
+        returnValueForMissingStub: _i9.Future<void>.value(),
+      ) as _i9.Future<void>);
   @override
   _i9.Future<void> setOnHttpError(_i3.HttpResponseErrorCallback? onHttpError) =>
       (super.noSuchMethod(
@@ -400,7 +400,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
-
   @override
   _i3.PlatformWebViewControllerCreationParams get params => (super.noSuchMethod(
         Invocation.getter(#params),
@@ -414,7 +413,6 @@ class MockAndroidWebViewController extends _i1.Mock
           Invocation.getter(#params),
         ),
       ) as _i3.PlatformWebViewControllerCreationParams);
-
   @override
   _i9.Future<void> loadFile(String? absoluteFilePath) => (super.noSuchMethod(
         Invocation.method(
@@ -424,7 +422,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> loadFlutterAsset(String? key) => (super.noSuchMethod(
         Invocation.method(
@@ -434,7 +431,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> loadHtmlString(
     String? html, {
@@ -449,7 +445,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> loadRequest(_i3.LoadRequestParams? params) =>
       (super.noSuchMethod(
@@ -460,7 +455,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<String?> currentUrl() => (super.noSuchMethod(
         Invocation.method(
@@ -470,7 +464,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<String?>.value(),
         returnValueForMissingStub: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
-
   @override
   _i9.Future<bool> canGoBack() => (super.noSuchMethod(
         Invocation.method(
@@ -480,7 +473,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<bool>.value(false),
         returnValueForMissingStub: _i9.Future<bool>.value(false),
       ) as _i9.Future<bool>);
-
   @override
   _i9.Future<bool> canGoForward() => (super.noSuchMethod(
         Invocation.method(
@@ -490,7 +482,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<bool>.value(false),
         returnValueForMissingStub: _i9.Future<bool>.value(false),
       ) as _i9.Future<bool>);
-
   @override
   _i9.Future<void> goBack() => (super.noSuchMethod(
         Invocation.method(
@@ -500,7 +491,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> goForward() => (super.noSuchMethod(
         Invocation.method(
@@ -510,7 +500,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> reload() => (super.noSuchMethod(
         Invocation.method(
@@ -520,7 +509,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> clearCache() => (super.noSuchMethod(
         Invocation.method(
@@ -530,7 +518,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> clearLocalStorage() => (super.noSuchMethod(
         Invocation.method(
@@ -540,7 +527,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setPlatformNavigationDelegate(
           _i3.PlatformNavigationDelegate? handler) =>
@@ -552,7 +538,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> runJavaScript(String? javaScript) => (super.noSuchMethod(
         Invocation.method(
@@ -562,7 +547,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<Object> runJavaScriptReturningResult(String? javaScript) =>
       (super.noSuchMethod(
@@ -585,7 +569,6 @@ class MockAndroidWebViewController extends _i1.Mock
           ),
         )),
       ) as _i9.Future<Object>);
-
   @override
   _i9.Future<void> addJavaScriptChannel(
           _i3.JavaScriptChannelParams? javaScriptChannelParams) =>
@@ -597,7 +580,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> removeJavaScriptChannel(String? javaScriptChannelName) =>
       (super.noSuchMethod(
@@ -608,7 +590,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<String?> getTitle() => (super.noSuchMethod(
         Invocation.method(
@@ -618,7 +599,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<String?>.value(),
         returnValueForMissingStub: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
-
   @override
   _i9.Future<void> scrollTo(
     int? x,
@@ -635,7 +615,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> scrollBy(
     int? x,
@@ -652,7 +631,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<_i4.Offset> getScrollPosition() => (super.noSuchMethod(
         Invocation.method(
@@ -674,7 +652,6 @@ class MockAndroidWebViewController extends _i1.Mock
           ),
         )),
       ) as _i9.Future<_i4.Offset>);
-
   @override
   _i9.Future<void> enableZoom(bool? enabled) => (super.noSuchMethod(
         Invocation.method(
@@ -684,7 +661,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setBackgroundColor(_i4.Color? color) => (super.noSuchMethod(
         Invocation.method(
@@ -694,7 +670,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setJavaScriptMode(_i3.JavaScriptMode? javaScriptMode) =>
       (super.noSuchMethod(
@@ -705,7 +680,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setUserAgent(String? userAgent) => (super.noSuchMethod(
         Invocation.method(
@@ -715,7 +689,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setMediaPlaybackRequiresUserGesture(bool? require) =>
       (super.noSuchMethod(
@@ -726,7 +699,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setTextZoom(int? textZoom) => (super.noSuchMethod(
         Invocation.method(
@@ -736,7 +708,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnShowFileSelector(
           _i9.Future<List<String>> Function(_i8.FileSelectorParams)?
@@ -749,7 +720,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnPlatformPermissionRequest(
           void Function(_i3.PlatformWebViewPermissionRequest)?
@@ -762,7 +732,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setGeolocationPermissionsPromptCallbacks({
     _i8.OnGeolocationPermissionsShowPrompt? onShowPrompt,
@@ -780,7 +749,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setCustomWidgetCallbacks({
     required _i8.OnShowCustomWidgetCallback? onShowCustomWidget,
@@ -798,7 +766,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setOnConsoleMessage(
           void Function(_i3.JavaScriptConsoleMessage)? onConsoleMessage) =>
@@ -810,7 +777,6 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<String?> getUserAgent() => (super.noSuchMethod(
         Invocation.method(
@@ -839,7 +805,6 @@ class MockAndroidWebViewProxy extends _i1.Mock
           Invocation.getter(#createAndroidWebView),
         ),
       ) as _i2.WebView Function());
-
   @override
   _i2.WebChromeClient Function({
     void Function(
@@ -968,7 +933,6 @@ class MockAndroidWebViewProxy extends _i1.Mock
           _i2.FileChooserParams,
         )? onShowFileChooser,
       }));
-
   @override
   _i2.WebViewClient Function({
     void Function(
@@ -990,6 +954,12 @@ class MockAndroidWebViewProxy extends _i1.Mock
       String,
       String,
     )? onReceivedError,
+    void Function(
+      _i2.WebView,
+      _i2.HttpAuthHandler,
+      String,
+      String,
+    )? onReceivedHttpAuthRequest,
     void Function(
       _i2.WebView,
       _i2.WebResourceRequest,
@@ -1025,6 +995,12 @@ class MockAndroidWebViewProxy extends _i1.Mock
             String,
             String,
           )? onReceivedError,
+          void Function(
+            _i2.WebView,
+            _i2.HttpAuthHandler,
+            String,
+            String,
+          )? onReceivedHttpAuthRequest,
           void Function(
             _i2.WebView,
             _i2.WebResourceRequest,
@@ -1065,6 +1041,12 @@ class MockAndroidWebViewProxy extends _i1.Mock
           )? onReceivedError,
           void Function(
             _i2.WebView,
+            _i2.HttpAuthHandler,
+            String,
+            String,
+          )? onReceivedHttpAuthRequest,
+          void Function(
+            _i2.WebView,
             _i2.WebResourceRequest,
             _i2.WebResourceError,
           )? onReceivedRequestError,
@@ -1103,6 +1085,12 @@ class MockAndroidWebViewProxy extends _i1.Mock
         )? onReceivedError,
         void Function(
           _i2.WebView,
+          _i2.HttpAuthHandler,
+          String,
+          String,
+        )? onReceivedHttpAuthRequest,
+        void Function(
+          _i2.WebView,
           _i2.WebResourceRequest,
           _i2.WebResourceError,
         )? onReceivedRequestError,
@@ -1115,7 +1103,6 @@ class MockAndroidWebViewProxy extends _i1.Mock
           String,
         )? urlLoading,
       }));
-
   @override
   _i2.FlutterAssetManager Function() get createFlutterAssetManager =>
       (super.noSuchMethod(
@@ -1129,7 +1116,6 @@ class MockAndroidWebViewProxy extends _i1.Mock
           Invocation.getter(#createFlutterAssetManager),
         ),
       ) as _i2.FlutterAssetManager Function());
-
   @override
   _i2.JavaScriptChannel Function(
     String, {
@@ -1156,7 +1142,6 @@ class MockAndroidWebViewProxy extends _i1.Mock
         String, {
         required void Function(String) postMessage,
       }));
-
   @override
   _i2.DownloadListener Function(
       {required void Function(
@@ -1199,7 +1184,6 @@ class MockAndroidWebViewProxy extends _i1.Mock
             String,
             int,
           ) onDownloadStart}));
-
   @override
   _i9.Future<void> setWebContentsDebuggingEnabled(bool? enabled) =>
       (super.noSuchMethod(
@@ -1230,7 +1214,6 @@ class MockAndroidWebViewWidgetCreationParams extends _i1.Mock
           Invocation.getter(#instanceManager),
         ),
       ) as _i5.InstanceManager);
-
   @override
   _i6.PlatformViewsServiceProxy get platformViewsServiceProxy =>
       (super.noSuchMethod(
@@ -1244,14 +1227,12 @@ class MockAndroidWebViewWidgetCreationParams extends _i1.Mock
           Invocation.getter(#platformViewsServiceProxy),
         ),
       ) as _i6.PlatformViewsServiceProxy);
-
   @override
   bool get displayWithHybridComposition => (super.noSuchMethod(
         Invocation.getter(#displayWithHybridComposition),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   _i3.PlatformWebViewController get controller => (super.noSuchMethod(
         Invocation.getter(#controller),
@@ -1264,14 +1245,12 @@ class MockAndroidWebViewWidgetCreationParams extends _i1.Mock
           Invocation.getter(#controller),
         ),
       ) as _i3.PlatformWebViewController);
-
   @override
   _i4.TextDirection get layoutDirection => (super.noSuchMethod(
         Invocation.getter(#layoutDirection),
         returnValue: _i4.TextDirection.rtl,
         returnValueForMissingStub: _i4.TextDirection.rtl,
       ) as _i4.TextDirection);
-
   @override
   Set<_i11.Factory<_i12.OneSequenceGestureRecognizer>> get gestureRecognizers =>
       (super.noSuchMethod(
@@ -1293,21 +1272,18 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   int get viewId => (super.noSuchMethod(
         Invocation.getter(#viewId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
-
   @override
   bool get awaitingCreation => (super.noSuchMethod(
         Invocation.getter(#awaitingCreation),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   _i7.PointTransformer get pointTransformer => (super.noSuchMethod(
         Invocation.getter(#pointTransformer),
@@ -1320,7 +1296,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
           Invocation.getter(#pointTransformer),
         ),
       ) as _i7.PointTransformer);
-
   @override
   set pointTransformer(_i7.PointTransformer? transformer) => super.noSuchMethod(
         Invocation.setter(
@@ -1329,14 +1304,12 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   bool get isCreated => (super.noSuchMethod(
         Invocation.getter(#isCreated),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   List<_i7.PlatformViewCreatedCallback> get createdCallbacks =>
       (super.noSuchMethod(
@@ -1344,7 +1317,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: <_i7.PlatformViewCreatedCallback>[],
         returnValueForMissingStub: <_i7.PlatformViewCreatedCallback>[],
       ) as List<_i7.PlatformViewCreatedCallback>);
-
   @override
   _i9.Future<void> setOffset(_i4.Offset? off) => (super.noSuchMethod(
         Invocation.method(
@@ -1354,7 +1326,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> create({
     _i4.Size? size,
@@ -1372,7 +1343,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<_i4.Size> setSize(_i4.Size? size) => (super.noSuchMethod(
         Invocation.method(
@@ -1394,7 +1364,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
           ),
         )),
       ) as _i9.Future<_i4.Size>);
-
   @override
   _i9.Future<void> sendMotionEvent(_i7.AndroidMotionEvent? event) =>
       (super.noSuchMethod(
@@ -1405,7 +1374,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   void addOnPlatformViewCreatedListener(
           _i7.PlatformViewCreatedCallback? listener) =>
@@ -1416,7 +1384,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   void removeOnPlatformViewCreatedListener(
           _i7.PlatformViewCreatedCallback? listener) =>
@@ -1427,7 +1394,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   _i9.Future<void> setLayoutDirection(_i4.TextDirection? layoutDirection) =>
       (super.noSuchMethod(
@@ -1438,7 +1404,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> dispatchPointerEvent(_i12.PointerEvent? event) =>
       (super.noSuchMethod(
@@ -1449,7 +1414,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> clearFocus() => (super.noSuchMethod(
         Invocation.method(
@@ -1459,7 +1423,6 @@ class MockExpensiveAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> dispose() => (super.noSuchMethod(
         Invocation.method(
@@ -1485,7 +1448,6 @@ class MockFlutterAssetManager extends _i1.Mock
         returnValue: _i9.Future<List<String?>>.value(<String?>[]),
         returnValueForMissingStub: _i9.Future<List<String?>>.value(<String?>[]),
       ) as _i9.Future<List<String?>>);
-
   @override
   _i9.Future<String> getAssetFilePathByName(String? name) =>
       (super.noSuchMethod(
@@ -1527,14 +1489,12 @@ class MockJavaScriptChannel extends _i1.Mock implements _i2.JavaScriptChannel {
           Invocation.getter(#channelName),
         ),
       ) as String);
-
   @override
   void Function(String) get postMessage => (super.noSuchMethod(
         Invocation.getter(#postMessage),
         returnValue: (String message) {},
         returnValueForMissingStub: (String message) {},
       ) as void Function(String));
-
   @override
   _i2.JavaScriptChannel copy() => (super.noSuchMethod(
         Invocation.method(
@@ -1568,7 +1528,6 @@ class MockPermissionRequest extends _i1.Mock implements _i2.PermissionRequest {
         returnValue: <String>[],
         returnValueForMissingStub: <String>[],
       ) as List<String>);
-
   @override
   _i9.Future<void> grant(List<String>? resources) => (super.noSuchMethod(
         Invocation.method(
@@ -1578,7 +1537,6 @@ class MockPermissionRequest extends _i1.Mock implements _i2.PermissionRequest {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> deny() => (super.noSuchMethod(
         Invocation.method(
@@ -1588,7 +1546,6 @@ class MockPermissionRequest extends _i1.Mock implements _i2.PermissionRequest {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i2.PermissionRequest copy() => (super.noSuchMethod(
         Invocation.method(
@@ -1671,7 +1628,6 @@ class MockPlatformViewsServiceProxy extends _i1.Mock
           ),
         ),
       ) as _i7.ExpensiveAndroidViewController);
-
   @override
   _i7.SurfaceAndroidViewController initSurfaceAndroidView({
     required int? id,
@@ -1738,21 +1694,18 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   int get viewId => (super.noSuchMethod(
         Invocation.getter(#viewId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
-
   @override
   bool get awaitingCreation => (super.noSuchMethod(
         Invocation.getter(#awaitingCreation),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   _i7.PointTransformer get pointTransformer => (super.noSuchMethod(
         Invocation.getter(#pointTransformer),
@@ -1765,7 +1718,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
           Invocation.getter(#pointTransformer),
         ),
       ) as _i7.PointTransformer);
-
   @override
   set pointTransformer(_i7.PointTransformer? transformer) => super.noSuchMethod(
         Invocation.setter(
@@ -1774,14 +1726,12 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   bool get isCreated => (super.noSuchMethod(
         Invocation.getter(#isCreated),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
-
   @override
   List<_i7.PlatformViewCreatedCallback> get createdCallbacks =>
       (super.noSuchMethod(
@@ -1789,7 +1739,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: <_i7.PlatformViewCreatedCallback>[],
         returnValueForMissingStub: <_i7.PlatformViewCreatedCallback>[],
       ) as List<_i7.PlatformViewCreatedCallback>);
-
   @override
   _i9.Future<void> setOffset(_i4.Offset? off) => (super.noSuchMethod(
         Invocation.method(
@@ -1799,7 +1748,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> create({
     _i4.Size? size,
@@ -1817,7 +1765,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<_i4.Size> setSize(_i4.Size? size) => (super.noSuchMethod(
         Invocation.method(
@@ -1839,7 +1786,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
           ),
         )),
       ) as _i9.Future<_i4.Size>);
-
   @override
   _i9.Future<void> sendMotionEvent(_i7.AndroidMotionEvent? event) =>
       (super.noSuchMethod(
@@ -1850,7 +1796,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   void addOnPlatformViewCreatedListener(
           _i7.PlatformViewCreatedCallback? listener) =>
@@ -1861,7 +1806,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   void removeOnPlatformViewCreatedListener(
           _i7.PlatformViewCreatedCallback? listener) =>
@@ -1872,7 +1816,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   _i9.Future<void> setLayoutDirection(_i4.TextDirection? layoutDirection) =>
       (super.noSuchMethod(
@@ -1883,7 +1826,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> dispatchPointerEvent(_i12.PointerEvent? event) =>
       (super.noSuchMethod(
@@ -1894,7 +1836,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> clearFocus() => (super.noSuchMethod(
         Invocation.method(
@@ -1904,7 +1845,6 @@ class MockSurfaceAndroidViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> dispose() => (super.noSuchMethod(
         Invocation.method(
@@ -1930,7 +1870,6 @@ class MockWebChromeClient extends _i1.Mock implements _i2.WebChromeClient {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setSynchronousReturnValueForOnConsoleMessage(bool? value) =>
       (super.noSuchMethod(
@@ -1941,7 +1880,6 @@ class MockWebChromeClient extends _i1.Mock implements _i2.WebChromeClient {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i2.WebChromeClient copy() => (super.noSuchMethod(
         Invocation.method(
@@ -1978,7 +1916,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setJavaScriptCanOpenWindowsAutomatically(bool? flag) =>
       (super.noSuchMethod(
@@ -1989,7 +1926,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setSupportMultipleWindows(bool? support) =>
       (super.noSuchMethod(
@@ -2000,7 +1936,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setJavaScriptEnabled(bool? flag) => (super.noSuchMethod(
         Invocation.method(
@@ -2010,7 +1945,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setUserAgentString(String? userAgentString) =>
       (super.noSuchMethod(
@@ -2021,7 +1955,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setMediaPlaybackRequiresUserGesture(bool? require) =>
       (super.noSuchMethod(
@@ -2032,7 +1965,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setSupportZoom(bool? support) => (super.noSuchMethod(
         Invocation.method(
@@ -2042,7 +1974,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setLoadWithOverviewMode(bool? overview) =>
       (super.noSuchMethod(
@@ -2053,7 +1984,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setUseWideViewPort(bool? use) => (super.noSuchMethod(
         Invocation.method(
@@ -2063,7 +1993,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setDisplayZoomControls(bool? enabled) => (super.noSuchMethod(
         Invocation.method(
@@ -2073,7 +2002,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setBuiltInZoomControls(bool? enabled) => (super.noSuchMethod(
         Invocation.method(
@@ -2083,7 +2011,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setAllowFileAccess(bool? enabled) => (super.noSuchMethod(
         Invocation.method(
@@ -2093,7 +2020,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setTextZoom(int? textZoom) => (super.noSuchMethod(
         Invocation.method(
@@ -2103,7 +2029,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<String> getUserAgentString() => (super.noSuchMethod(
         Invocation.method(
@@ -2126,7 +2051,6 @@ class MockWebSettings extends _i1.Mock implements _i2.WebSettings {
           ),
         )),
       ) as _i9.Future<String>);
-
   @override
   _i2.WebSettings copy() => (super.noSuchMethod(
         Invocation.method(
@@ -2166,7 +2090,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
           Invocation.getter(#settings),
         ),
       ) as _i2.WebSettings);
-
   @override
   _i9.Future<void> loadData({
     required String? data,
@@ -2186,7 +2109,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> loadDataWithBaseUrl({
     String? baseUrl,
@@ -2210,7 +2132,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> loadUrl(
     String? url,
@@ -2227,7 +2148,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> postUrl(
     String? url,
@@ -2244,7 +2164,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<String?> getUrl() => (super.noSuchMethod(
         Invocation.method(
@@ -2254,7 +2173,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<String?>.value(),
         returnValueForMissingStub: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
-
   @override
   _i9.Future<bool> canGoBack() => (super.noSuchMethod(
         Invocation.method(
@@ -2264,7 +2182,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<bool>.value(false),
         returnValueForMissingStub: _i9.Future<bool>.value(false),
       ) as _i9.Future<bool>);
-
   @override
   _i9.Future<bool> canGoForward() => (super.noSuchMethod(
         Invocation.method(
@@ -2274,7 +2191,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<bool>.value(false),
         returnValueForMissingStub: _i9.Future<bool>.value(false),
       ) as _i9.Future<bool>);
-
   @override
   _i9.Future<void> goBack() => (super.noSuchMethod(
         Invocation.method(
@@ -2284,7 +2200,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> goForward() => (super.noSuchMethod(
         Invocation.method(
@@ -2294,7 +2209,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> reload() => (super.noSuchMethod(
         Invocation.method(
@@ -2304,7 +2218,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> clearCache(bool? includeDiskFiles) => (super.noSuchMethod(
         Invocation.method(
@@ -2314,7 +2227,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<String?> evaluateJavascript(String? javascriptString) =>
       (super.noSuchMethod(
@@ -2325,7 +2237,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<String?>.value(),
         returnValueForMissingStub: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
-
   @override
   _i9.Future<String?> getTitle() => (super.noSuchMethod(
         Invocation.method(
@@ -2335,7 +2246,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<String?>.value(),
         returnValueForMissingStub: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
-
   @override
   _i9.Future<void> scrollTo(
     int? x,
@@ -2352,7 +2262,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> scrollBy(
     int? x,
@@ -2369,7 +2278,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<int> getScrollX() => (super.noSuchMethod(
         Invocation.method(
@@ -2379,7 +2287,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<int>.value(0),
         returnValueForMissingStub: _i9.Future<int>.value(0),
       ) as _i9.Future<int>);
-
   @override
   _i9.Future<int> getScrollY() => (super.noSuchMethod(
         Invocation.method(
@@ -2389,7 +2296,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<int>.value(0),
         returnValueForMissingStub: _i9.Future<int>.value(0),
       ) as _i9.Future<int>);
-
   @override
   _i9.Future<_i4.Offset> getScrollPosition() => (super.noSuchMethod(
         Invocation.method(
@@ -2411,7 +2317,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
           ),
         )),
       ) as _i9.Future<_i4.Offset>);
-
   @override
   _i9.Future<void> setWebViewClient(_i2.WebViewClient? webViewClient) =>
       (super.noSuchMethod(
@@ -2422,7 +2327,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> addJavaScriptChannel(
           _i2.JavaScriptChannel? javaScriptChannel) =>
@@ -2434,7 +2338,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> removeJavaScriptChannel(
           _i2.JavaScriptChannel? javaScriptChannel) =>
@@ -2446,7 +2349,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setDownloadListener(_i2.DownloadListener? listener) =>
       (super.noSuchMethod(
@@ -2457,7 +2359,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setWebChromeClient(_i2.WebChromeClient? client) =>
       (super.noSuchMethod(
@@ -2468,7 +2369,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i9.Future<void> setBackgroundColor(_i4.Color? color) => (super.noSuchMethod(
         Invocation.method(
@@ -2478,7 +2378,6 @@ class MockWebView extends _i1.Mock implements _i2.WebView {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i2.WebView copy() => (super.noSuchMethod(
         Invocation.method(
@@ -2517,7 +2416,6 @@ class MockWebViewClient extends _i1.Mock implements _i2.WebViewClient {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i2.WebViewClient copy() => (super.noSuchMethod(
         Invocation.method(
@@ -2554,7 +2452,6 @@ class MockWebStorage extends _i1.Mock implements _i2.WebStorage {
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
-
   @override
   _i2.WebStorage copy() => (super.noSuchMethod(
         Invocation.method(
@@ -2588,7 +2485,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         returnValue: (int __p0) {},
         returnValueForMissingStub: (int __p0) {},
       ) as void Function(int));
-
   @override
   set onWeakReferenceRemoved(void Function(int)? _onWeakReferenceRemoved) =>
       super.noSuchMethod(
@@ -2598,7 +2494,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   int addDartCreatedInstance(_i5.Copyable? instance) => (super.noSuchMethod(
         Invocation.method(
@@ -2608,7 +2503,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
-
   @override
   int? removeWeakReference(_i5.Copyable? instance) => (super.noSuchMethod(
         Invocation.method(
@@ -2617,7 +2511,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         ),
         returnValueForMissingStub: null,
       ) as int?);
-
   @override
   T? remove<T extends _i5.Copyable>(int? identifier) => (super.noSuchMethod(
         Invocation.method(
@@ -2626,7 +2519,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         ),
         returnValueForMissingStub: null,
       ) as T?);
-
   @override
   T? getInstanceWithWeakReference<T extends _i5.Copyable>(int? identifier) =>
       (super.noSuchMethod(
@@ -2636,7 +2528,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         ),
         returnValueForMissingStub: null,
       ) as T?);
-
   @override
   int? getIdentifier(_i5.Copyable? instance) => (super.noSuchMethod(
         Invocation.method(
@@ -2645,7 +2536,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         ),
         returnValueForMissingStub: null,
       ) as int?);
-
   @override
   void addHostCreatedInstance(
     _i5.Copyable? instance,
@@ -2661,7 +2551,6 @@ class MockInstanceManager extends _i1.Mock implements _i5.InstanceManager {
         ),
         returnValueForMissingStub: null,
       );
-
   @override
   bool containsIdentifier(int? identifier) => (super.noSuchMethod(
         Invocation.method(
